# control/src/main/java/app/collide/control/auth/dto/LoginRequest.java
package app.collide.control.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

/** Email + password login. No strong-password check here — that's a signup concern. */
public record LoginRequest(
        @NotBlank @Email @Size(max = 254) String email,
        @NotBlank @Size(max = 128) String password) {}